Fix use of uninitialized memory in simple_tf2_core tests.
Since geometry_msgs::msg::TransformStamped is declared with
an empty constructor, all of its fields are uninitialized.
This means that anytime we want to use it, we must initialize
all fields ourselves.  The tests in simple_tf2_core.cpp were
failing to do this, so parts of tf2 were making decisions on
basically random data.  This was causing a bunch of valgrind
failures.  With this commit, the test becomes valgrind clean.

Signed-off-by: Chris Lalancette <clalancette@osrfoundation.org>