Fix high CPU - Use executor to spin and stop node in tf_listener thread (#119)
Previously, #114 resolved the issue of terminating the work in the 
thread spinning the node which lead to hanging and unclean shutdown when 
destructing the node. However, the current implementation runs in a
tight loop using rclcpp::spin_some, which leads to very high and
inefficient cpu usage even when there is no work to be processed.

This problem was resolved in ros-planning/navigation2#866, by putting 
the node on an executor and spinning. As opposed to rclcpp::spin, this
change allows the executor to manage the termination of the spin thread
by calling executor->cancel in the destructor. This pattern seems to
work well when needing to cleanly join a thread spinning a node.

Signed-off-by: bpwilcox <bpwilcox@eng.ucsd.edu>