Change TF2Error names to be a bit more descriptive.
The driving reason behind this is to avoid the NO_ERROR macro
on Windows.  We can't totally get around the problem right now,
but we can deprecate the old names and switch to new ones.
After Galactic, we can remove the old names and get rid of the
problem completely.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>