Covariance transformation now accepts tf Transform
Previously, `transformCovariance`'s transform argument was of type
`geometry_msgs::msg::Transform`, which is not a "general" transform
type, but rather a message type. As such, that argument type has been
changed to `tf2::Transform`, so that the user can convert from whatever
transform type they have to a tf Transform.

Also:

* Moved location of a forward declaration to be closer to its use.
* Applied `ament_uncrustify` changes for maximum line length, changing
  the `transformCovariance` signature ine.
* Reflected change in `transformCovariance` function signature in
  doxygen documentation.

Signed-off-by: Abrar Rahman Protyasha <abrar@openrobotics.org>