Switch to a single mutex for all data.
This makes it much easier to reason about what is going to happen
in here, and actually protects all of the data.  We had to
make it a recursive mutex because we want user callbacks to
be able to call back into this code.  Unfortunately, that means
we can't use thread annotation since clang doesn't support
thread annotation with recursive mutexes.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>