Disable mem-access warnings on aarch64.
There is currently a bug in Eigen in Ubuntu 22.04 meaning
that it throws a mem-access warning every time Eigen/PacketMath.h
is used.  There is an upstream fix for it, and we are trying
to get that into Ubuntu 22.04, but that is not guaranteed.
For now, disable the warning only if we are on GCC or Clang
and the target architecture is aarch64.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>