Port point cloud transformation to numpy (#507)
Implements the point cloud transformation in NumPy. This speeds
up the transformation operation by 20000% for large point clouds
(excluding the serialization and deserialization). The code needs
some further testing but should work as it is. I also try to improve
the performance of the serialization and deserialization as seen in
https://github.com/ros2/common_interfaces/pull/175.

This adds NumPy as a dependency and drops the dependency on
PyKDL which should solve #471.

The API is also improved by adding proper interface descriptions.

Signed-off-by: Florian Vahl <florian@flova.de>
Co-authored-by: Abrar Rahman Protyasha <aprotyas@u.rochester.edu>