Fix hard-coded Windows paths.
These come about because of the use of the old-style CMake
variables like ${orocos_kdl_INCLUDE_DIRS}.  Unfortunately,
there is a bug in the orocos_kdl CMake modern targets where
they forget to export a dependency on Eigen.  To workaround
this, make an explicit dependency on Eigen (even though we
don't directly use it), and use the Eigen targets instead.

Note that on Windows, this *still* causes us to use a
hardcoded path on Eigen.  However, this path is less problematic
because it will always be there if the user followed our
installation instructions (and won't be a "random" path like
C:\ci\ws\install\include).  Also, once we fix
https://github.com/ros2/choco-packages/issues/19 , this hard-coded
path will go away as well.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>