Fix use-after-free bug in BufferServer::cancelCB (#579)
`info` was a reference to the `GoalInfo` entry in `active_goals_` that
`it` pointed at. It was used after the entry had been erased from the
list (through `it`). Swapping the line with `erase` and the one with the
call to `canceled` would have been the minimal fix, but instead we do a
refactor here to improve the overall code quality and make such errors
less likely in the future.

Signed-off-by: Alexander Hans <ahans@users.noreply.github.com>